Change persist to merge on classes with embedded id. Change findById and existsById from int to long in DAOs. Remove some model classes validations (caused DB conflicts)

+++ src/main/java/co/edu/icesi/researchgroupmanagement/daos/interfaces/AdmissionRequestDao.java
@@ -10,6 +10,6 @@ public interface AdmissionRequestDao {
 	public AdmissionRequest update(AdmissionRequest entity);
 	public void delete (AdmissionRequest entity);
 	public List<AdmissionRequest> findAll();
-	public AdmissionRequest findById(Integer id);
-	public boolean existsById(Integer id);
+	public AdmissionRequest findById(Long id);
+	public boolean existsById(Long id);
 }
