+++ src/main/java/co/edu/icesi/researchgroupmanagement/daos/interfaces/RprojectStatusDao.java
@@ -10,6 +10,6 @@ public interface RprojectStatusDao {
 	public RprojectStatus update(RprojectStatus entity);
 	public void delete (RprojectStatus entity);
 	public List<RprojectStatus> findAll();
-	public RprojectStatus findById(Integer id);
-	public boolean existsById(Integer id);
+	public RprojectStatus findById(Long id);
+	public boolean existsById(Long id);
 }
